fix EndpointVariantError introduced on 2024-05-28 by botocore endpoint update
botocore commit 2a598d0c5c from 2024-05-28 added the new eu-isoe-west-1
region to the aws-iso-e partition. This is actually the first region to
be defined in endpoints.json that belongs to an ISO partition *not*
flagged as "dualstack not supported"; the other ISO partitions are
either flagged (aws-iso, aws-iso-b) or do not define any regions.

We will assume that the aws-iso-e partition should also be flagged as
well, as none of the services seem to support dualstack.

Therefore,
- add aws-iso-e to botocore's _UNSUPPORTED_DUALSTACK_PARTITIONS
- add eu-isoe-west-1 as test case
- intercept and log EndpointVariantError